+++ Tests/CreationTestCase.swift
@@ -37,12 +37,12 @@ class CreationTestCase: XCTestCase {
         let first = MemoryStorage()
         let second = MemoryStorage()
         controller.manager.storage = first
-        expect(first.delegate === controller.manager).to(beTrue())
+        expect(first.delegate === controller.manager.collectionViewUpdater).to(beTrue())
         
         controller.manager.storage = second
         
         expect(first.delegate == nil).to(beTrue())
-        expect(second.delegate === controller.manager).to(beTrue())
+        expect(second.delegate === controller.manager.collectionViewUpdater).to(beTrue())
     }
     
     func testCreatingCollectionControllerFromCode()